# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Dimohod Trade API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CategoryNode {
  id: string;
  /** @nullable */
  parent_id?: string | null;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  sort_order: number;
  children: CategoryNode[];
}

export interface CatalogTreeResponse {
  items: CategoryNode[];
}

export interface CategorySummary {
  id: string;
  name: string;
  slug: string;
}

export interface Sku {
  id: string;
  article: string;
  name: string;
  /** @nullable */
  price_rub?: string | null;
  stock_status: string;
}

export interface Product {
  id: string;
  category: CategorySummary;
  name: string;
  slug: string;
  /** @nullable */
  short_description?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  wall_thickness_mm?: string | null;
  /** @nullable */
  diameter_mm?: number | null;
  application_tags: string[];
  /** @nullable */
  compatibility_notes?: string | null;
  skus: Sku[];
}