# lib/api-zod/src/generated/types/catalogTreeResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Dimohod Trade API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryNode } from './categoryNode';

export interface CatalogTreeResponse {
  items: CategoryNode[];
}