# lib/api-zod/src/generated/types/categoryNode.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Dimohod Trade API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryNode {
  id: string;
  /** @nullable */
  parent_id?: string | null;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  sort_order: number;
  children: CategoryNode[];
}