# lib/api-zod/src/generated/types/categorySummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Dimohod Trade API
 * OpenAPI spec version: 0.1.0
 */

export interface CategorySummary {
  id: string;
  name: string;
  slug: string;
}