# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Dimohod Trade API
 * OpenAPI spec version: 0.1.0
 */
import type { CategorySummary } from './categorySummary';
import type { Sku } from './sku';

export interface Product {
  id: string;
  category: CategorySummary;
  name: string;
  slug: string;
  /** @nullable */
  short_description?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  wall_thickness_mm?: string | null;
  /** @nullable */
  diameter_mm?: number | null;
  application_tags: string[];
  /** @nullable */
  compatibility_notes?: string | null;
  skus: Sku[];
}